# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs_api
  contact: {}
  version: '1.0'
...
servers:
  - url: localhost:3000/api/v1
    variables: {}
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: GET all jobs
      operationId: GETalljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    post:
      tags:
        - Jobs
      summary: POST job
      operationId: POSTjob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/POSTjobRequest'
                - examples:
                    - company: test
                      position: posiiton
              contentMediaType: application/json
            example:
              company: test
              position: posiiton
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    parameters: []
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
        - Jobs
      summary: GET unique job
      operationId: GETuniquejob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    delete:
      tags:
        - Jobs
      summary: DELETE job
      operationId: DELETEjob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    patch:
      tags:
        - Jobs
      summary: PATCH job
      operationId: PATCHjob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PATCHjobRequest'
                - examples:
                    - company: jhon company
                      position: fullstack dev
              contentMediaType: application/json
            example:
              company: jhon company
              position: fullstack dev
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    parameters: []
components:
  schemas:
...
    POSTjobRequest:
      title: POSTjobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      examples:
        - company: test
          position: posiiton
    PATCHjobRequest:
      title: PATCHjobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      examples:
        - company: jhon company
          position: fullstack dev
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...